Cast int to be correct

+++ i386/LineRenderer.c
@@ -129,7 +129,7 @@ void terminal_backspace(void)
 	write_serial('\b');
 	write_serial(' ');
 	write_serial('\b');
-	if (terminal_column == -1) {
+	if (terminal_column == (size_t)-1) {
 		// ^[1A - Move cursor one line up
 		write_serial('\033');
 		write_serial('[');
